Refactor block overview aggregation
